# openapi-eventia.yaml
openapi: 3.0.1
info:
  title: Eventia API
  description: API REST para la gestión de eventos, artistas, recintos, usuarios y reservas.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
...
paths:
...
  /reservas:
    get:
      tags: [Reservas]
      summary: Obtener todas las reservas
      description: Filtra opcionalmente por metodoPago, codigoReserva y confirmada.
      parameters:
        - name: metodoPago
          in: query
          required: false
          description: Filtra por método de pago.
          schema:
            type: string
          example: tarjeta
        - name: codigoReserva
          in: query
          required: false
          description: Filtra por código de reserva.
          schema:
            type: string
          example: RSV-2026-001
        - name: confirmada
          in: query
          required: false
          description: Filtra por si la reserva está confirmada.
          schema:
            type: boolean
          example: true
      responses:
        '200':
          description: Lista de reservas obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reserva'
              examples:
                sinFiltros:
                  summary: Consulta sin filtros
                  value:
                    - id: 1
                      fechaReserva: '2026-04-01T18:30:00'
                      cantidadEntradas: 2
                      precioTotal: 91.0
                      metodoPago: tarjeta
                      codigoReserva: RSV-2026-001
                      confirmada: true
                      usuario:
                        id: 1
                      evento:
                        id: 1
                    - id: 2
                      fechaReserva: '2026-04-05T10:00:00'
                      cantidadEntradas: 4
                      precioTotal: 100.0
                      metodoPago: bizum
                      codigoReserva: RSV-2026-002
                      confirmada: false
                      usuario:
                        id: 2
                      evento:
                        id: 2
                filtradoPorMetodoPago:
                  summary: Consulta filtrada por método de pago tarjeta
                  value:
                    - id: 1
                      fechaReserva: '2026-04-01T18:30:00'
                      cantidadEntradas: 2
                      precioTotal: 91.0
                      metodoPago: tarjeta
                      codigoReserva: RSV-2026-001
                      confirmada: true
                      usuario:
                        id: 1
                      evento:
                        id: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Reservas]
      summary: Crear reserva
      description: Registra una nueva reserva.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reserva'
            examples:
              ejemplo1:
                summary: Reserva con tarjeta
                value:
                  fechaReserva: '2026-04-01T18:30:00'
                  cantidadEntradas: 2
                  precioTotal: 91.0
                  metodoPago: tarjeta
                  codigoReserva: RSV-2026-001
                  confirmada: true
                  usuario:
                    id: 1
                  evento:
                    id: 1
              ejemplo2:
                summary: Reserva con Bizum
                value:
                  fechaReserva: '2026-04-05T10:00:00'
                  cantidadEntradas: 4
                  precioTotal: 100.0
                  metodoPago: bizum
                  codigoReserva: RSV-2026-002
                  confirmada: false
                  usuario:
                    id: 2
                  evento:
                    id: 2
      responses:
        '201':
          description: Reserva creada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
              examples:
                ejemplo1:
                  summary: Reserva con tarjeta creada
                  value:
                    id: 1
                    fechaReserva: '2026-04-01T18:30:00'
                    cantidadEntradas: 2
                    precioTotal: 91.0
                    metodoPago: tarjeta
                    codigoReserva: RSV-2026-001
                    confirmada: true
                    usuario:
                      id: 1
                    evento:
                      id: 1
                ejemplo2:
                  summary: Reserva con Bizum creada
                  value:
                    id: 2
                    fechaReserva: '2026-04-05T10:00:00'
                    cantidadEntradas: 4
                    precioTotal: 100.0
                    metodoPago: bizum
                    codigoReserva: RSV-2026-002
                    confirmada: false
                    usuario:
                      id: 2
                    evento:
                      id: 2
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reservas/{id}:
    get:
      tags: [Reservas]
      summary: Obtener reserva por id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Reserva obtenida correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
              examples:
                ejemplo1:
                  summary: Reserva confirmada
                  value:
                    id: 1
                    fechaReserva: '2026-04-01T18:30:00'
                    cantidadEntradas: 2
                    precioTotal: 91.0
                    metodoPago: tarjeta
                    codigoReserva: RSV-2026-001
                    confirmada: true
                    usuario:
                      id: 1
                    evento:
                      id: 1
                ejemplo2:
                  summary: Reserva pendiente
                  value:
                    id: 2
                    fechaReserva: '2026-04-05T10:00:00'
                    cantidadEntradas: 4
                    precioTotal: 100.0
                    metodoPago: bizum
                    codigoReserva: RSV-2026-002
                    confirmada: false
                    usuario:
                      id: 2
                    evento:
                      id: 2
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Reservas]
      summary: Modificar reserva
      description: Sustituye completamente los datos de una reserva existente.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reserva'
            examples:
              ejemplo1:
                summary: Confirmar reserva
                value:
                  fechaReserva: '2026-04-05T10:00:00'
                  cantidadEntradas: 4
                  precioTotal: 100.0
                  metodoPago: bizum
                  codigoReserva: RSV-2026-002
                  confirmada: true
                  usuario:
                    id: 2
                  evento:
                    id: 2
              ejemplo2:
                summary: Cambiar método de pago
                value:
                  fechaReserva: '2026-04-01T18:30:00'
                  cantidadEntradas: 2
                  precioTotal: 91.0
                  metodoPago: efectivo
                  codigoReserva: RSV-2026-001
                  confirmada: true
                  usuario:
                    id: 1
                  evento:
                    id: 1
      responses:
        '200':
          description: Reserva modificada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
              examples:
                ejemplo1:
                  summary: Reserva confirmada
                  value:
                    id: 2
                    fechaReserva: '2026-04-05T10:00:00'
                    cantidadEntradas: 4
                    precioTotal: 100.0
                    metodoPago: bizum
                    codigoReserva: RSV-2026-002
                    confirmada: true
                    usuario:
                      id: 2
                    evento:
                      id: 2
                ejemplo2:
                  summary: Método de pago actualizado
                  value:
                    id: 1
                    fechaReserva: '2026-04-01T18:30:00'
                    cantidadEntradas: 2
                    precioTotal: 91.0
                    metodoPago: efectivo
                    codigoReserva: RSV-2026-001
                    confirmada: true
                    usuario:
                      id: 1
                    evento:
                      id: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Reservas]
      summary: Modificar parcialmente una reserva
      description: Actualiza uno o varios campos de una reserva existente.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            examples:
              ejemplo1:
                summary: Desconfirmar reserva
                value:
                  confirmada: false
              ejemplo2:
                summary: Cambiar método de pago
                value:
                  metodoPago: bizum
      responses:
        '200':
          description: Reserva actualizada parcialmente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
              examples:
                ejemplo1:
                  summary: Reserva desconfirmada
                  value:
                    id: 1
                    fechaReserva: '2026-04-01T18:30:00'
                    cantidadEntradas: 2
                    precioTotal: 91.0
                    metodoPago: tarjeta
                    codigoReserva: RSV-2026-001
                    confirmada: false
                    usuario:
                      id: 1
                    evento:
                      id: 1
                ejemplo2:
                  summary: Reserva con método de pago actualizado
                  value:
                    id: 2
                    fechaReserva: '2026-04-05T10:00:00'
                    cantidadEntradas: 4
                    precioTotal: 100.0
                    metodoPago: bizum
                    codigoReserva: RSV-2026-002
                    confirmada: false
                    usuario:
                      id: 2
                    evento:
                      id: 2
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Reservas]
      summary: Eliminar reserva
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Reserva eliminada correctamente
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reservas/confirmadas:
    get:
      tags: [Reservas]
      summary: Obtener reservas confirmadas
      description: Devuelve las reservas con confirmada = true mediante una consulta JPQL.
      responses:
        '200':
          description: Lista de reservas confirmadas obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reserva'
              examples:
                ejemplo1:
                  summary: Una reserva confirmada
                  value:
                    - id: 1
                      fechaReserva: '2026-04-01T18:30:00'
                      cantidadEntradas: 2
                      precioTotal: 91.0
                      metodoPago: tarjeta
                      codigoReserva: RSV-2026-001
                      confirmada: true
                      usuario:
                        id: 1
                      evento:
                        id: 1
                ejemplo2:
                  summary: Varias reservas confirmadas
                  value:
                    - id: 1
                      fechaReserva: '2026-04-01T18:30:00'
                      cantidadEntradas: 2
                      precioTotal: 91.0
                      metodoPago: tarjeta
                      codigoReserva: RSV-2026-001
                      confirmada: true
                      usuario:
                        id: 1
                      evento:
                        id: 1
                    - id: 3
                      fechaReserva: '2026-04-10T09:00:00'
                      cantidadEntradas: 1
                      precioTotal: 45.5
                      metodoPago: efectivo
                      codigoReserva: RSV-2026-003
                      confirmada: true
                      usuario:
                        id: 3
                      evento:
                        id: 1
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Identificador único del recurso
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  responses:
    BadRequest:
      description: Error de validación o petición incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            title: bad-request
            message: Error de validación
            errors:
              email: El email no es válido
              password: La contraseña debe tener entre 6 y 20 caracteres
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            title: not-found
            message: Recurso no encontrado
            errors: {}
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            title: internal-server-error
            message: Error interno del servidor
            errors: {}
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        title:
          type: string
          example: bad-request
        message:
          type: string
          example: Error de validación
        errors:
          type: object
          additionalProperties:
            type: string
      required:
        - code
        - title
        - message
        - errors
    UsuarioRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
      required:
        - id
...
    EventoRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
      required:
        - id
...
    Reserva:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        fechaReserva:
          type: string
          format: date-time
          example: '2026-04-01T18:30:00'
        cantidadEntradas:
          type: integer
          minimum: 1
          example: 2
        precioTotal:
          type: number
          format: float
          minimum: 0
          example: 91.0
        metodoPago:
          type: string
          example: tarjeta
        codigoReserva:
          type: string
          minLength: 3
          example: RSV-2026-001
        confirmada:
          type: boolean
          example: true
        usuario:
          $ref: '#/components/schemas/UsuarioRef'
        evento:
          $ref: '#/components/schemas/EventoRef'
      required:
        - fechaReserva
        - cantidadEntradas
        - precioTotal
        - codigoReserva
        - usuario
        - evento